Cancelar Agendamento

+++ site/frontend/src/pages/ListaAgenda.js
@@ -33,13 +33,39 @@ const ListaAgenda = () => {
     }
 
     async function deletar(agendaR) {
-        const id = agendaR.id;
-        try {
-            await api.delete(`/doacao/${id}`);
-            fetchItens();
-        } catch(err) {
-            swal("Erro", "Erro ao deletar", "error");
-        }
+        swal("Atenção","Deseja Excluir?", {
+            closeOnClickOutside: false, //não fecha o swal se clicar fora
+            dangerMode: true, //danger mode pra chamar atenção
+            closeOnEsc: false, // não deixa fechar no esc
+            buttons: {
+                cancelar: {  // Posso mudar o nome aqui
+                    text: "Cancelar", // texto do botão
+                    value: "cancelar", // valor pra gente testar la em baixo
+                    className: "swal-button--danger", // classe do botão css
+                },
+                confirmar: { // botao confirmar
+                    text: "Confirmar", // texto do botão
+                    value: "confirmar", // valor pra gente testar la em baixo
+                    className: "swal-button--confirm", // classe do botão css
+                },
+
+            },
+            icon: "info", // icone do swal
+        })
+        .then(async (value) => { // aqui é a ação !
+            if(value == "confirmar"){ // se clicou em "confirmar" chama o ajax que faz a exclusão.
+                const id = agendaR.doacao_id;
+                try {
+                    await api.delete(`/doacao/${id}`);
+                    swal("Excluido!", "Exclusão efetuada com sucesso.", "success").then(function() {
+                        window.location = '/listaDoacao';
+                    });
+                    // fetchItens();
+                } catch(err) {
+                    swal("Erro", "Erro ao deletar", "error");
+                }
+            }
+        });
     }
 
     function alterar(ItensR) {
